+++ tools/shaders_glsl/hunterk/handheld/console-border/shader-files/gb-pass-4.glsl
@@ -66,8 +66,8 @@ COMPAT_VARYING vec4 TEX0;
 COMPAT_VARYING vec4 TEX1;
  
 uniform mat4 MVPMatrix;
-uniform int FrameDirection;
-uniform int FrameCount;
+uniform float FrameDirection;
+uniform float FrameCount;
 uniform COMPAT_PRECISION vec2 OutputSize;
 uniform COMPAT_PRECISION vec2 TextureSize;
 uniform COMPAT_PRECISION vec2 InputSize;
@@ -143,8 +143,8 @@ COMPAT_VARYING vec4 TEX1;
  
 uniform sampler2D Texture;
 uniform sampler2D Pass2Texture;
-uniform int FrameDirection;
-uniform int FrameCount;
+uniform float FrameDirection;
+uniform float FrameCount;
 uniform COMPAT_PRECISION vec2 OutputSize;
 uniform COMPAT_PRECISION vec2 TextureSize;
 uniform COMPAT_PRECISION vec2 InputSize;
